add MaybeObjectTypeTrait
