+++ src/Type/IterableIterableType.php
@@ -2,14 +2,13 @@
 
 namespace PHPStan\Type;
 
-use PHPStan\Analyser\Scope;
-use PHPStan\Reflection\ClassConstantReflection;
-use PHPStan\Reflection\MethodReflection;
-use PHPStan\Reflection\PropertyReflection;
 use PHPStan\TrinaryLogic;
+use PHPStan\Type\Traits\MaybeObjectTypeTrait;
 
 class IterableIterableType implements StaticResolvableType, CompoundType
 {
+
+	use MaybeObjectTypeTrait;
 
 	/** @var \PHPStan\Type\Type */
 	private $keyType;
@@ -140,56 +139,6 @@ class IterableIterableType implements StaticResolvableType, CompoundType
 	public function isCallable(): TrinaryLogic
 	{
 		return TrinaryLogic::createMaybe();
-	}
-
-	public function canAccessProperties(): bool
-	{
-		return false;
-	}
-
-	public function hasProperty(string $propertyName): bool
-	{
-		return false;
-	}
-
-	public function getProperty(string $propertyName, Scope $scope): PropertyReflection
-	{
-		throw new \PHPStan\ShouldNotHappenException();
-	}
-
-	public function canCallMethods(): bool
-	{
-		return false;
-	}
-
-	public function hasMethod(string $methodName): bool
-	{
-		return false;
-	}
-
-	public function getMethod(string $methodName, Scope $scope): MethodReflection
-	{
-		throw new \PHPStan\ShouldNotHappenException();
-	}
-
-	public function canAccessConstants(): bool
-	{
-		return false;
-	}
-
-	public function hasConstant(string $constantName): bool
-	{
-		return false;
-	}
-
-	public function getConstant(string $constantName): ClassConstantReflection
-	{
-		throw new \PHPStan\ShouldNotHappenException();
-	}
-
-	public function isClonable(): bool
-	{
-		return false;
 	}
 
 	public static function __set_state(array $properties): Type
